+++ packages/xy/tests/scatter/scatter.props.ts
@@ -1,5 +1,4 @@
 import dataPolynomials from './dataPolynomials.json'
-import dataDistributions from './dataDistributions.json'
 
 export const scatterProps = {
     data: dataPolynomials,
@@ -40,16 +39,5 @@ export const timeScatterProps = {
         variant: 'linear' as const,
         min: 0,
         max: 64,
-    },
-}
-
-export const histogramProps = {
-    data: dataDistributions,
-    breaks: 4,
-    scaleX: {
-        variant: 'linear' as const,
-    },
-    scaleY: {
-        variant: 'linear' as const,
     },
 }