+++ packages/xy/tests/scatter/signals.test.tsx
@@ -1,5 +1,5 @@
-import { convolution, curvePoints, downsample } from '../src/scatter'
 import { roundDecimalPlaces } from '@chsk/core'
+import { convolution, curvePoints, downsample } from '../../src/scatter'
 
 const round2dp = (x: number) => roundDecimalPlaces(x, 2)